pyproject.toml
# SPDX-License-Identifier: MPL-2.0 AND LicenseRef-Commons-Clause-License-Condition-1.0
# <!-- // /*  d a r k s h a p e s */ -->

[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
authors = [{ name = "darkshapes", email = "91800957+exdysa@users.noreply.github.com" }]
name = "negate"
description = "Synthetic image detector"
readme = "README.md"
dynamic = ["version"]
license = { file = "LICENSE" }
requires-python = ">=3.13"
keywords = ["AI", "ML", "GenAI", "training", "torch"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "accelerate>=1.12.0",
    "altair>=6.1.0.dev20260126",
    "datasets>=4.5.0",
    "huggingface-hub>=1.3.2",
    "ipython>=9.10.0",
    "mongoengine>=0.29.1",
    "numpy==2.4.1",
    "open-clip-torch>=3.2.0",
    "pillow>=12.1.0",
    "seaborn>=0.13.2",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.57.3",
]
[dependency-groups]
dev = ["pyright>=1.1.408", "pytest>=9.0.2", "pytest-asyncio>=1.3.0", "ruff>=0.14.13"]
[project.optional-dependencies]
# -- frameworks --

# -- accelerators --
cuda = ["torch", "torchvision"]
rocm = ["torch", "torchvision"]
cpu = ["torch", "torchvision"]
mps = ["torch", "torchvision"]

[project.urls]
Homepage = "https://github.com/darkshapes/negate"

[project.scripts]
negate = "negate.__main__:main"

[tool.uv]
torch-backend = "auto"
prerelease = "allow"
preview = true

conflicts = [[{ extra = "cpu" }, { extra = "mps" }, { extra = "rocm" }, { extra = "cuda" }]]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]

torchvision = [{ index = "pytorch-cpu", extra = "mps", marker = "sys_platform == 'darwin'" }]

torch = [{ index = "pytorch-cpu", extra = "mps", marker = "sys_platform == 'darwin'" }]

[tool.setuptools_scm]
write_to = "_version.py"

[tool.setuptools.packages.find]
where = ["."]
include = ["negate*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
negate = ["negate/config/*.yaml", "negate/results/*"]

[tool.ruff]
line-length = 140
include = ["*.py"]
extend-exclude = ["^tests/.*$", "test.*$"]

[tool.ruff.lint]
ignore = ["E731"]

[tool.ruff.lint.pycodestyle]
max-line-length = 140
ignore-overlong-task-comments = true

[tool.pylint]
max-line-length = 140
ignore-paths = ["^tests/.*$", "test_.*$"]
disable = ["C0415"]

[tool.pyright]
include = ["negate"]
exclude = [
    "^tests/.*$",
    "**/__pycache__", # cache directories
    "./typings",      # generated type stubs
]

[tool.typos]
files.extend-exclude = ["^tests/.*$", "test.*$"]
default.extend-words = { "ot" = "ot", "convers" = "convers" }
